requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pymupdf==1.24.9
tika==2.6.0
elasticsearch==7.17.9
pydantic==2.8.2
python-dotenv==1.0.1